Add BepInEx data to analytics embed
